chore: refactor underlying data provider
